refactor: mock function testUtils로 이동
